# metal : use F16 math in mul_mat kernels