--- a/ggml/src/ggml-metal/ggml-metal.m
+++ b/ggml/src/ggml-metal/ggml-metal.m
@@ -2021,7 +2021,8 @@ static void ggml_metal_encode_node(
                     [encoder setBuffer:id_src1 offset:offs_src1    atIndex:2];
                     [encoder setBuffer:id_dst  offset:offs_dst     atIndex:3];
 
-                    [encoder setThreadgroupMemoryLength:8192 atIndex:0];
+                    [encoder setThreadgroupMemoryLength:8192        atIndex:0];
+                  //[encoder setThreadgroupMemoryLength:4096 + 2048 atIndex:0];
                     [encoder dispatchThreadgroups:MTLSizeMake( (ne11 + 31)/32, (ne01 + 63)/64, ne12*ne13) threadsPerThreadgroup:MTLSizeMake(128, 1, 1)];
                 } else {
                     int nth0 = 32;