Using F16 math and accumulators in matrix multiplication kernels.

The prompt processing performance is improved. The perplexity is affected as follows:

```bash
./llama-perplexity -m models/llama-3.1-8b/ggml-model-f16.gguf -ngl 99 -f build/wikitext-2-raw/wiki.test.raw
```

```
master: PPL = 6.4007 +/- 0.03938
PR:     PPL = 6.4385 +/- 0.03969
```

Not sure if it worth it yet. Might do this as a compile-time flag `GGML_METAL_F16`.

```bash
./scripts/compare-commits.sh master gg/metal-mul-mat-f16 -m ./models/llama-3.2-3b-instruct/ggml-model-f16.gguf -m ./models/llama-3.2-3b-instruct/ggml-model-q8_0.gguf -m ./models/llama-3.2-1b-instruct/ggml-model-f16.gguf -m ./models/llama-3.2-1b-instruct/ggml-model-q8_0.gguf -m models/qwen2.5-7b-coder/ggml-model-q8_0.gguf -m models/qwen2.5-1.5b-coder/ggml-model-q8_0.gguf -m models/mistral-7b-v0.2/ggml-model-q8_0.gguf -m models/gemma-2b/ggml-model-q4_0.gguf -m models/gemma-2b/ggml-model-f16.gguf -fa 1 -p 512,4096 -b 4096 -ub 4096 -n 0
```

| CPU                        | Model         | Test   |   t/s master |   t/s gg/metal-mul-mat-f16 |   Speedup |
|:---------------------------|:--------------|:-------|-------------:|---------------------------:|----------:|
| Accelerate, Apple M2 Ultra | gemma 2B F16  | pp512  |      4340.17 |                    4487.07 |      1.03 |
| Accelerate, Apple M2 Ultra | gemma 2B F16  | pp4096 |      3924.03 |                    3993.83 |      1.02 |
| Accelerate, Apple M2 Ultra | gemma 2B Q4_0 | pp512  |      3905.70 |                    4159.25 |      1.06 |
| Accelerate, Apple M2 Ultra | gemma 2B Q4_0 | pp4096 |      3547.61 |                    3754.79 |      1.06 |
| Accelerate, Apple M2 Ultra | llama 1B F16  | pp512  |      8627.93 |                    8838.55 |      1.02 |
| Accelerate, Apple M2 Ultra | llama 1B F16  | pp4096 |      7799.88 |                    7867.90 |      1.01 |
| Accelerate, Apple M2 Ultra | llama 1B Q8_0 | pp512  |      7797.25 |                    8262.59 |      1.06 |
| Accelerate, Apple M2 Ultra | llama 1B Q8_0 | pp4096 |      7193.80 |                    7457.12 |      1.04 |
| Accelerate, Apple M2 Ultra | llama 3B F16  | pp512  |      3234.86 |                    3293.91 |      1.02 |
| Accelerate, Apple M2 Ultra | llama 3B F16  | pp4096 |      3089.56 |                    3131.40 |      1.01 |
| Accelerate, Apple M2 Ultra | llama 3B Q8_0 | pp512  |      2938.62 |                    3082.19 |      1.05 |
| Accelerate, Apple M2 Ultra | llama 3B Q8_0 | pp4096 |      2817.66 |                    2935.23 |      1.04 |
| Accelerate, Apple M2 Ultra | llama 7B Q8_0 | pp512  |      1246.76 |                    1362.89 |      1.09 |
| Accelerate, Apple M2 Ultra | llama 7B Q8_0 | pp4096 |      1231.98 |                    1328.87 |      1.08 |
| Accelerate, Apple M2 Ultra | qwen2 ?B Q8_0 | pp512  |      3494.22 |                    3719.73 |      1.06 |
| Accelerate, Apple M2 Ultra | qwen2 ?B Q8_0 | pp4096 |      3411.72 |                    3634.29 |      1.07 |