@joseph777111 2024-12-09
A compile time flag seems like a great idea! With that, I think it would nice to have to flags: 

1. GGML_METAL_F16
2. GGML_METAL_F32

That way the User can choose the precision. 😋