@ggerganov 2024-12-12
It will be checked at run-time, based on the precision configured for each matrix multiplication (see https://github.com/ggerganov/llama.cpp/pull/10665#issuecomment-2519524632).